refactor: 댓글 및 사용자 상태 관리 최적화와 관련된 컴포넌트 추가 Manager 분리
